oops forgot this

+++ src/integral/ecp/soangularbatch.cc
@@ -37,7 +37,7 @@ const static SphUSPList sphusplist;
 const static DoubleFactorial df;
 
 SOBatch::SOBatch(const std::shared_ptr<const SOECP> _so, const std::array<std::shared_ptr<const Shell>,2>& _info,
-                           const double contA, const double contC, const std::array<int, 3> angA, const std::array<int, 3> angC,
+                           const int contA, const int contC, const std::array<int, 3> angA, const std::array<int, 3> angC,
                            const bool print, const int max_iter, const double thresh_int)
  : RadialInt(3, print, max_iter, thresh_int),
    basisinfo_(_info), so_(_so), cont0_(contA), cont1_(contC), ang0_(angA), ang1_(angC) {
